+++ lib/octodown.rb
@@ -1,10 +1,10 @@
 require 'octodown/renderer/github_markdown'
 require 'octodown/renderer/html'
 require 'octodown/support/services/document_presenter'
+require 'octodown/support/services/riposter'
 require 'octodown/support/relative_root_filter'
 require 'octodown/support/html_file'
 require 'octodown/support/server'
-require 'octodown/support/riposter'
 require 'octodown/version'
 
 module Octodown
